requirements.txt
streamlit
numpy
pandas
pillow
plotly
requests
scikit-learn
reportlab
joblib
easyocr
tensorflow